pyproject.toml
[tool.poetry]
name = "warmit"
version = "0.1.0"
description = "AI-powered email warming tool to improve deliverability"
authors = ["Giovanni Rapa"]
readme = "README.md"
packages = [{include = "warmit", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
pydantic = "^2.10.0"
pydantic-settings = "^2.6.0"
aiosmtplib = "^3.0.2"
aioimap = "^1.1.1"
python-whois = "^0.9.4"
dnspython = "^2.7.0"
openai = "^1.58.1"
apscheduler = "^3.10.4"
python-dotenv = "^1.0.1"
asyncpg = "^0.30.0"
email-validator = "^2.2.0"
httpx = "^0.28.1"
redis = "^5.2.0"
celery = "^5.4.0"
typer = "^0.15.1"
rich = "^13.9.4"
aiosqlite = "^0.20.0"
streamlit = "^1.41.1"
plotly = "^5.24.1"
pandas = "^2.2.3"
psutil = "^6.1.1"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.24.0"
black = "^24.10.0"
ruff = "^0.8.4"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
